增加蓝图
