+++ .idea/CSG_MonitoringSystem.iml
@@ -14,7 +14,7 @@
     <option name="TEMPLATE_CONFIGURATION" value="Jinja2" />
     <option name="TEMPLATE_FOLDERS">
       <list>
-        <option value="$MODULE_DIR$/../dynamicDemo01\templates" />
+        <option value="$MODULE_DIR$/templates" />
       </list>
     </option>
   </component>
