+++ apps/controller/__init__.py
@@ -1,0 +1,2 @@
+# -*- codeing = utf-8 -*-
+
