+++ apps/settings.py
@@ -1,0 +1,2 @@
+ENV = 'development'
+DEBUG = True
