+++ requirement.txt
@@ -5,8 +5,6 @@ importlib-metadata==4.5.0
 itsdangerous==2.0.1
 Jinja2==3.0.1
 MarkupSafe==2.0.1
-six==1.16.0
-tornado==6.1
 typing-extensions==3.10.0.0
 Werkzeug==2.0.1
 zipp==3.4.1
