+++ static/css/monitorStyle.css
@@ -103,7 +103,6 @@ option {
 pre {
     white-space: pre-wrap;
     white-space: -moz-pre-wrap;
-    white-space: -pre-wrap;
     white-space: -o-pre-wrap;
     word-wrap: break-word;
 }
