+++ static/js/controller_ajax/controller_monitor.js
