+++ static/js/controller_ajax/controller_suricata.js
