+++ templates/flow.html
@@ -106,7 +106,7 @@
 <script type="text/javascript" src="static/js/flow/flowTimeRank.js"></script>
 <script type="text/javascript" src="static/js/flow/equipmentFlowRank.js"></script>
 <script type="text/javascript" src="static/js/flow/queryAssets.js"></script>
-<script type="text/javascript" src="../static/js/controller_flow.js"></script>
+<script type="text/javascript" src="../static/js/controller_ajax/controller_flow.js"></script>
 <script type="text/javascript">
     $('document').ready(function () {
         $("body").css('visibility', 'visible');
