+++ templates/monitor.html
@@ -99,7 +99,7 @@
 <script type="text/javascript" src="../static/js/monitor/terminal_label_percentage_04.js"></script>
 <script type="text/javascript" src="../static/js/monitor/terminal_label_percentage_05.js"></script>
 <script type="text/javascript" src="../static/js/monitor/terminal_label_percentage_06.js"></script>
-<script type="text/javascript" src="../static/js/controller_monitor.js"></script>
+<script type="text/javascript" src="../static/js/controller_ajax/controller_monitor.js"></script>
