+++ templates/suricata.html
@@ -122,7 +122,7 @@
 <script type="text/javascript" src="../static/js/suricata/suricata_flowTrend.js"></script>
 <script type="text/javascript" src="../static/js/suricata/suricata_AbnormalFlow.js"></script>
 <script type="text/javascript" src="../static/js/suricata/suricata_attackDisplay.js"></script>
-<script type="text/javascript" src="../static/js/controller_suricata.js"></script>
+<script type="text/javascript" src="../static/js/controller_ajax/controller_suricata.js"></script>
 
 <script type="text/javascript">
     $('document').ready(function () {
